auto_doc_markdown_converter/src/docx_extractor.py
# Placeholder for DOCX extraction logic 

import docx

def extract_text_from_docx(file_path: str) -> str:
    """
    从 .docx 文件中提取所有文本，保留段落分隔。

    参数:
        file_path: .docx 文件的路径。

    返回:
        包含提取文本的字符串，段落之间用换行符分隔。
    """
    doc = docx.Document(file_path)
    full_text = []
    for para in doc.paragraphs:
        full_text.append(para.text)
    return "\n".join(full_text)

# 示例用法 (仅用于测试目的):
# if __name__ == '__main__':
#     # 创建一个临时的 docx 文件用于测试
#     from docx import Document
#     document = Document()
#     document.add_heading('文档标题', 0)
#     document.add_paragraph('这是第一个段落。')
#     document.add_paragraph('这是第二个段落，包含更多文本。')
#     document.add_paragraph('以及第三个段落。')
#     dummy_docx_path = "sample_for_extractor.docx"
#     document.save(dummy_docx_path)
#
#     try:
#         text = extract_text_from_docx(dummy_docx_path)
#         print(f"从 {dummy_docx_path} 提取的文本:")
#         print(text)
#     except Exception as e:
#         print(f"发生错误: {e}")
#     finally:
#         # 清理临时文件
#         import os
#         if os.path.exists(dummy_docx_path):
#             os.remove(dummy_docx_path) 