requirements.txt
python-docx
pdfplumber
reportlab 